+++ CHANGES.md
@@ -2,7 +2,8 @@
 * none 
 
 **Changed:**
-* none
+* moved some release automation things around
+* removed .build file and instead storing it in jar manifest
 
 **Fixed:**
 * none